x86/set_memory: skip vm_unmap_aliases for PKS (un)protect
Calling vm_unmap_aliases() on pks_unprotect path could create a deadlock
because vm_unmap_aliases() releases page tables with in turn calls
pks_unprotect again. This deadlock is not reached with grouped cache
because actual freeing of the page tables is postponed to shrinker, but
switching to a caching integrated with page allocator would hit the
deadlock.

The page table pages are never mapped in vmalloc area, so they won't have
an alias there, hence it is safe to omit the call to vm_unmap_aliases()
when removing PKS protection from the page tables.

Signed-off-by: Mike Rapoport <rppt@linux.ibm.com>