x86/sev: add a SVSM vTPM platform device
If the SNP boot has a SVSM, probe for the vTPM device by sending a
call to SVSM function 8 with no arguments.  If this returns
successfully, the vTPM is present in the SVSM (If the SVSM doesn't
have a vTPM, this call should return SVSM_ERR_UNSUPPORTED_CALLID).

If a vTPM is found, register a platform device as "platform:tpm" so it
can be attached to the tpm_platform.c driver.

Signed-off-by: James Bottomley <James.Bottomley@HansenPartnership.com>