mm, treewide: make get_free_pages() and friends return void *
Vast majority of allocations that use get_free_pages() and it's
derivatives cast the returned value to a pointer type.

Make get_free_pages() and it derivatives return 'void *' with the exception of __get_dma_pages() whose callers all expect an unsigned long.

The call sites are updated with the semantic patch below and a bit of
manual tweaking.

    @@
    type T;
    T addr;
    expression gfp, order;
    @@
    (
    - addr = __get_free_pages(gfp, order)
    + addr = (T)__get_free_pages(gfp, order)
    |
    - addr = __get_free_page(gfp)
    + addr = (T)__get_free_page(gfp)
    |
    - addr = get_zeroed_page(gfp)
    + addr = (T)get_zeroed_page(gfp)
    )

    @@
    type T;
    @@
    (
    - (T *)
      \(__get_free_pages\|get_zeroed_page\|__get_free_page\)(...)
    )

Signed-off-by: Mike Rapoport (Microsoft) <rppt@kernel.org>